requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
requests==2.32.3
redis==5.2.1
slowapi==0.1.9


